Audit and polish: fix uint16 texture-size truncation and dead vars
- Widen tex_byte_size from uint16 to uint32 in CPVRTexture::LoadBMP and
  scherzo_bmp_load_texture. For any texture where width*height*2 exceeds
  65535 bytes (e.g. a 256x128 RGB565 background), the byte count wrapped
  around and pvr_mem_malloc under-allocated, corrupting PVR memory when
  the texture rows were copied in.
- Remove unused outer bufsrc/bufdst/x/y/cnt declarations in
  scherzo_sep_data; both branches use their own locals or memcpy.

+++ src/dc_utils.cpp
@@ -323,7 +323,7 @@ bool scherzo_bmp_load_texture(const char* filename, pvr_ptr_t* tex, uint32 *w, u
 		++h_count;
 	*tex_width = (int32) powf(2, w_count);
 	*tex_height = (int32) powf(2, h_count);
-	uint16 tex_byte_size = sizeof(uint16) * (*tex_width) * (*tex_height);
+	uint32 tex_byte_size = sizeof(uint16) * (*tex_width) * (*tex_height);
 	*tex = pvr_mem_malloc(tex_byte_size);
 	
 	uint16* temp_dest = (uint16*) *tex;
